vit_model_server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
opencv-python==4.7.0.72
numpy==1.24.3
matplotlib==3.7.1
tqdm==4.65.0
seaborn==0.12.2
scikit-learn==1.2.2
